Improve Documentation (#23)
Reorganize the main Readme so that it emphasis the most important information.

Move discussion of TypeTrees to a dedicated topic, and expand out to discussion Player and AssetBundle builds.

Add more examples and some missing arguments to the UnityDataTool command line topic. Reorganize so that the experimental reference command is listed last.

Fix a few errors and typos (and hopefully not introduce a few new ones)