Analysis Usability Improvements
Try to reduce the noise when files that are not SerializedFiles are found by Analysis.
Ignore files by name and extension that are almost certainly NOT SerializedFiles or Archives, which which are found in build output (Player, AssetBundle or ContentArchives)
Add a verbose mode and hide callstack and other info unless verbose is set.   When verbose is on then any ignored files are printed.